% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_D3partitionR.R
\name{add_data}
\alias{add_data}
\title{Append data to a D3partitionR object}
\usage{
add_data(D3partitionR_object, data, steps, count = "value", color = "name",
  label = "name", tooltip = "name", aggregate_fun = NULL)
}
\arguments{
\item{D3partitionR_object}{The D3partitionR object to which the data should be appended}

\item{data}{a data.frame object}

\item{steps}{The vector of steps to be used}

\item{count}{The variable to be used as the count variable, typically, the number of occurences.}

\item{color}{a variable to use as color (default: name)}

\item{label}{a variable to use as label (default: name)}

\item{tooltip}{a variable to use as tooltip (default: name)}

\item{aggregate_fun}{A named list of function which will be used to aggregates to variables used in color, label or tooltips. This only applies to variable in the provided dataset.}
}
\value{
The D3partitionR object with the appended data
}
\description{
Append data to a D3partitionR object
}
