% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_D3partitionR.R
\name{add_nodes_data}
\alias{add_nodes_data}
\title{Add informations (for instance new names, colors, ....) to the nodes of a D3_partitionR object}
\usage{
add_nodes_data(D3partitionR_object, nodes_data)
}
\arguments{
\item{D3partitionR_object}{The D3partitionR object to which the data should be appended}

\item{nodes_data}{a names list where the name of each element is the name of a node. The data will be appended to the node in the nested list}
}
\value{
The D3partitionR object with the appended nodes data
}
\description{
Add informations (for instance new names, colors, ....) to the nodes of a D3_partitionR object
}
