% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_D3partitionR.R
\name{add_title}
\alias{add_title}
\title{Add a title to a D3partitionR object}
\usage{
add_title(D3partitionR_object, text, style = NULL)
}
\arguments{
\item{D3partitionR_object}{The D3partitionR object to which the data should be appended}

\item{text}{Title text}

\item{style}{A valid CSS string which will be applied to the title)}
}
\value{
A D3partitionR object
}
\description{
Add a title to a D3partitionR object
}
