% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_D3partitionR.R
\name{compile_D3_partitionR}
\alias{compile_D3_partitionR}
\title{Compile D3partitionR object to plot it}
\usage{
compile_D3_partitionR(D3partitionR_object)
}
\arguments{
\item{D3partitionR_object}{The D3partitionR object to which the data should be appended}
}
\value{
A D3partitionR compiled object
}
\description{
Compile D3partitionR object to plot it
}
