% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_to_nest.R
\name{compute_unique_leaf_name}
\alias{compute_unique_leaf_name}
\title{Return al the leaf names}
\usage{
compute_unique_leaf_name(nested_list)
}
\arguments{
\item{nested_list}{A nested_list where each node has a name attribute}
}
\description{
Return al the leaf names
}
