% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_D3partitionR.R
\name{css_to_object}
\alias{css_to_object}
\title{Change a css string to a css 'object' to be used with d3.styles}
\usage{
css_to_object(css_string)
}
\arguments{
\item{css_string}{the string to be transformed}
}
\value{
A list of attribute
}
\description{
Change a css string to a css 'object' to be used with d3.styles
}
