% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_to_nest.R
\name{find_min_max_tree}
\alias{find_min_max_tree}
\title{Find the maximum values of a given var in a tree}
\usage{
find_min_max_tree(nested_list, variable = "value")
}
\arguments{
\item{nested_list}{A nested_list where each node has a name attribute}

\item{variable}{A nested_list where each node has a name attribute}
}
\description{
Find the maximum values of a given var in a tree
}
