% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_to_nest.R
\name{get_all_nodes_names}
\alias{get_all_nodes_names}
\title{Return al the possible nodes names}
\usage{
get_all_nodes_names(nested_list, variable = "name")
}
\arguments{
\item{nested_list}{A nested_list where each node has a name attribute}

\item{variable}{the variable to collect}
}
\description{
Return al the possible nodes names
}
