% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_D3partitionR.R
\name{is_present_variable}
\alias{is_present_variable}
\title{Check if a variable is present in a D3partitionR object}
\usage{
is_present_variable(variable, D3partitionR_object)
}
\arguments{
\item{variable}{The variable which presence is to be checked}

\item{D3partitionR_object}{The D3partitionR object}
}
\value{
TRUE/FALSE
}
\description{
Check if a variable is present in a D3partitionR object
}
