% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3partitionR.R
\name{plot.D3partitionR}
\alias{plot.D3partitionR}
\title{Plot D3partitionR object}
\usage{
\method{plot}{D3partitionR}(x, width = NULL, height = NULL,
  elementId = NULL, sizingPolicy = NULL, ...)
}
\arguments{
\item{x}{A D3partitionR object to plot}

\item{width}{width of the widget in pixel/percent}

\item{height}{height of the widget in pixel/percent}

\item{elementId}{html id of the widget}

\item{sizingPolicy}{sizing policy}

\item{...}{parameters for method consistency}
}
\description{
Plot D3partitionR object
}
