% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_D3partitionR.R
\name{scale_type}
\alias{scale_type}
\title{Check if the scale variable is discrete or continuous}
\usage{
scale_type(color_variable, D3partitionR_object)
}
\arguments{
\item{color_variable}{The color variable to be assessed}

\item{D3partitionR_object}{The D3partitionR object}
}
\value{
TRUE/FALSE
}
\description{
Check if the scale variable is discrete or continuous
}
