% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_D3partitionR.R
\name{set_chart_type}
\alias{set_chart_type}
\title{Set the chart_type}
\usage{
set_chart_type(D3partitionR_object, chart_type)
}
\arguments{
\item{D3partitionR_object}{The D3partitionR object to which the data should be appended}

\item{chart_type}{type fo chart to use (in c('sunburst','treemap','circle_treemap','partition_chart','icicle') )}
}
\value{
A D3partitionR object
}
\description{
Set the chart_type
}
