% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_D3partitionR.R
\name{set_continuous_color_scale}
\alias{set_continuous_color_scale}
\title{Add a custom discrete color scale}
\usage{
set_continuous_color_scale(D3partitionR_object, color_palette)
}
\arguments{
\item{D3partitionR_object}{The D3partitionR object to which the data should be appended}

\item{color_palette}{a vector of two colors, the first one is use on the bottom of the scale, the other on the top.}
}
\value{
A D3partitionR object
}
\description{
Add a custom discrete color scale
}
