% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_D3partitionR.R
\name{set_discrete_color_scale}
\alias{set_discrete_color_scale}
\title{Add a custom discrete color scale}
\usage{
set_discrete_color_scale(D3partitionR_object, color_palette)
}
\arguments{
\item{D3partitionR_object}{The D3partitionR object to which the data should be appended}

\item{color_palette}{A vector (or a named vector with levels of the variable color)}
}
\value{
A D3partitionR object
}
\description{
Add a custom discrete color scale
}
