% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_D3partitionR.R
\name{set_labels_parameters}
\alias{set_labels_parameters}
\title{Set the labels parameters}
\usage{
set_labels_parameters(D3partitionR_object, visible = T, cut_off = 3,
  style = NULL)
}
\arguments{
\item{D3partitionR_object}{The D3partitionR object to which the data should be appended}

\item{visible}{boolean, should the labels be diplayed ? Default: TRUE}

\item{cut_off}{a numeric variable between 0 and 100. Nodes which represent less than cut_off percents of the current root will have their labels hidden.}

\item{style}{a valid CSS string to be applied to the labels. Default: NULL}
}
\value{
A D3partitionR object
}
\description{
Set the labels parameters
}
