% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_D3partitionR.R
\name{set_legend_parameters}
\alias{set_legend_parameters}
\title{Set the legend parameter}
\usage{
set_legend_parameters(D3partitionR_object, visible = T, zoom_subset = F,
  width = 100)
}
\arguments{
\item{D3partitionR_object}{The D3partitionR object to which the data should be appended}

\item{visible}{boolean, should the trail be diplayed ? Default: TRUE}

\item{zoom_subset}{boolean, if TRUE, only the modalities present in the children of the zoomed rrot are displayed kin the legend.}

\item{width}{legend width in pixel}
}
\value{
A D3partitionR object
}
\description{
Set the legend parameter
}
