% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_D3partitionR.R
\name{set_nodes_styles}
\alias{set_nodes_styles}
\title{Set the node style}
\usage{
set_nodes_styles(D3partitionR_object,
  idle_style = "stroke : black; stroke-width : 2;",
  hovered_style = "stroke : white; stroke-width : 2;")
}
\arguments{
\item{D3partitionR_object}{The D3partitionR object to which the data should be appended}

\item{idle_style}{the CSS to be applied when the node is idle (i.e no click nor hover)}

\item{hovered_style}{the CSS to be applied when the node is hovered}
}
\value{
A D3partitionR object
}
\description{
Set the node style
}
