% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_D3partitionR.R
\name{set_shiny_input}
\alias{set_shiny_input}
\title{Configuration of a D3partitionR object as a Shiny input}
\usage{
set_shiny_input(D3partitionR_object, input_id,
  enabled_inputs = list(clicked_node = T, leaves = T, nodes = T, ancestors =
  T, children_path = F))
}
\arguments{
\item{D3partitionR_object}{The D3partitionR object to which the data should be appended}

\item{input_id}{The id of the input}

\item{enabled_inputs}{which inputs should be enabled ? defaut to list(clicked_node=T,leaf=T,nodes=T,ancestors=T,child_path=F)}
}
\value{
A D3partitionR object
}
\description{
Configuration of a D3partitionR object as a Shiny input
}
