% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_D3partitionR.R
\name{set_tooltip_parameters}
\alias{set_tooltip_parameters}
\title{Set the tooltips parameter}
\usage{
set_tooltip_parameters(D3partitionR_object, visible = T, style = NULL,
  builder = "table")
}
\arguments{
\item{D3partitionR_object}{The D3partitionR object to which the data should be appended}

\item{visible}{boolean, should the trail be diplayed ? Default: TRUE}

\item{style}{a valid CSS string to be applied to the tooltip. Default: NULL}

\item{builder}{Tooltip builder to use for the tooltip. Can either one of the predefined tooltip ('table','basic') or a js expression returning a tooltip.}
}
\value{
A D3partitionR object
}
\description{
Set the tooltips parameter
}
