% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_D3partitionR.R
\name{set_trail}
\alias{set_trail}
\title{Enable/disable the trail of steps}
\usage{
set_trail(D3partitionR_object, visible = T)
}
\arguments{
\item{D3partitionR_object}{The D3partitionR object to which the data should be appended}

\item{visible}{boolean, should the trail be diplayed ? Default: TRUE}
}
\value{
A D3partitionR object
}
\description{
Enable/disable the trail of steps
}
