% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/session_to_nest.R
\name{strip_path}
\alias{strip_path}
\title{Strip a dataframe containing a step into separate columns}
\usage{
strip_path(data, path_col = "path", count_col = "count",
  value_cols = NULL, sep = "->")
}
\arguments{
\item{data}{A dataframe containing the path.}

\item{path_col}{Name of the column containing the path. The path should be a string of the format "step 1 -> step 2 -> step 3" .Default: "path"}

\item{count_col}{Name of the column containing the number of occurences of the path. Default: "count"}

\item{value_cols}{Names of the other columns to keep. Default: NULL}

\item{sep}{String used to separate the different steps. Default: "->"}
}
\value{
A data.table with the columns specified in count_col, value_cols and one column per step in the path
}
\description{
Strip a dataframe containing a step into separate columns
}
