% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/builder.R
\name{tooltip_builder}
\alias{tooltip_builder}
\title{Build tooltip html function}
\usage{
tooltip_builder(type)
}
\arguments{
\item{type}{a tooltip type: 'basic' (i.e the variable value) or 'table'(i.e. a table with the variables names and value)}
}
\description{
Build tooltip html function
}
