% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/D3partitionR.R
\name{D3partitionR}
\alias{D3partitionR}
\title{D3partitionR}
\usage{
D3partitionR(random = F, data = NULL, type = "circleTreeMap",
  width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{random}{generates a radom CircleTreeMap, default to FALSE.}

\item{data}{The inputs data, it should be in one of the following form:
-a dataframe with two columns, the first one being the paths (and named path) and the second one the value of the paths.
-a list of two lists with the same properties.}

\item{width}{width of the element}

\item{height}{height of the element}

\item{elementId}{Id of the elemen, default to null}
}
\description{
This function generates the CircleTreeMap from your data or from random data
}

